[new] Middlewares faltantes del core

+++ sd-bank-integration-layer/src/integrations/core/middlewares/cajero.middleware.ts
@@ -13,6 +13,7 @@ import Usuario from "../../../shared/models/Usuario";
 import Perfil from "../../../shared/models/Perfil";
 import Sucursal from "../../../shared/models/Sucursal";
 import { Types } from "mongoose";
+import { Estado } from "../../../shared/utils/estado";
 
 // @desc     a new cashier
 // @route    POST
@@ -23,55 +24,64 @@ const createCajero = asyncHandler(
     res: Response,
     next: NextFunction
   ): Promise<void | Response> => {
-    const session: ClientSession = await startSession();
+    const estado = Estado.getInstance();
 
-    try {
-      session.startTransaction();
-      const { cedula, nombre, apellido, sucursal_id } = req.body;
+    if (estado.getArriba()) {
+      next();
+    } else {
+      const session: ClientSession = await startSession();
 
-      const cajeroEncontrado = await Cajero.findOne({ cedula }).session(
-        session
-      );
+      try {
+        session.startTransaction();
+        const { cedula, nombre, apellido, sucursal_id } = req.body;
 
-      if (cajeroEncontrado)
-        return next(
-          new ErrorResponse("Ya existe un cajero con esas credenciales.", 400)
+        const cajeroEncontrado = await Cajero.findOne({ cedula }).session(
+          session
         );
 
-      const sucursalEncontrada: any = await Sucursal.findById(
-        sucursal_id
-      ).session(session);
+        if (cajeroEncontrado)
+          return next(
+            new ErrorResponse("Ya existe un cajero con esas credenciales.", 400)
+          );
 
-      if (!sucursalEncontrada)
-        return notFound({ message: "No se halló la sucursal provista.", next });
+        const sucursalEncontrada: any = await Sucursal.findById(
+          sucursal_id
+        ).session(session);
 
-      const cajeroToCreate = {
-        cedula,
-        nombre,
-        apellido,
-        sucursal: sucursal_id as Types.ObjectId,
-      };
+        if (!sucursalEncontrada)
+          return notFound({
+            message: "No se halló la sucursal provista.",
+            next,
+          });
 
-      const cajeroCreado: any = await Cajero.create([cajeroToCreate], {
-        session,
-      });
+        const cajeroToCreate = {
+          cedula,
+          nombre,
+          apellido,
+          sucursal: sucursal_id as Types.ObjectId,
+        };
 
-      sucursalEncontrada.cajeros.push(cajeroCreado[0]._id as Types.ObjectId);
+        const cajeroCreado: any = await Cajero.create([cajeroToCreate], {
+          session,
+        });
 
-      await sucursalEncontrada.save();
+        sucursalEncontrada.cajeros.push(cajeroCreado[0]._id as Types.ObjectId);
 
-      await session.commitTransaction();
-      session.endSession();
+        await sucursalEncontrada.save();
 
-      res.status(200).json({
-        exito: true,
-        mensaje: "¡Cajero creado satisfactoriamente!",
-      });
-    } catch (err) {
-      await session.abortTransaction();
-      session.endSession();
+        await session.commitTransaction();
+        session.endSession();
 
-      return errorHandler(err, req, res, next);
+        res.status(200).json({
+          exito: true,
+          mensaje: "¡Cajero creado satisfactoriamente!",
+        });
+      } catch (err) {
+        await session.abortTransaction();
+        session.endSession();
+
+        return errorHandler(err, req, res, next);
+      }
     }
   }
 );
@@ -82,17 +92,29 @@ const getCajeroById = asyncHandler(
     res: Response,
     next: NextFunction
   ): Promise<void | Response> => {
-    const { _id } = req.params;
+    const estado = Estado.getInstance();
 
-    const cajero = await Cajero.findById(_id);
+    if (estado.getArriba()) {
+      next();
+    } else {
+      const estado = Estado.getInstance();
 
-    if (!cajero)
-      return notFound({
-        message: "No se halló ningún cajero con el _id provisto.",
-        next,
-      });
+      if (estado.getArriba()) {
+        next();
+      } else {
+        const { _id } = req.params;
 
-    res.status(200).json(cajero);
+        const cajero = await Cajero.findById(_id);
+
+        if (!cajero)
+          return notFound({
+            message: "No se halló ningún cajero con el _id provisto.",
+            next,
+          });
+
+        res.status(200).json(cajero);
+      }
+    }
   }
 );
 const getCajeroByCedula = asyncHandler(
@@ -101,17 +123,29 @@ const getCajeroByCedula = asyncHandler(
     res: Response,
     next: NextFunction
   ): Promise<void | Response> => {
-    const { cedula } = req.params;
+    const estado = Estado.getInstance();
 
-    const cajero = await Cajero.findOne({ cedula });
+    if (estado.getArriba()) {
+      next();
+    } else {
+      const estado = Estado.getInstance();
 
-    if (!cajero)
-      return notFound({
-        message: "No se halló ningún cajero con la cédula provista.",
-        next,
-      });
+      if (estado.getArriba()) {
+        next();
+      } else {
+        const { cedula } = req.params;
 
-    res.status(200).json(cajero);
+        const cajero = await Cajero.findOne({ cedula });
+
+        if (!cajero)
+          return notFound({
+            message: "No se halló ningún cajero con la cédula provista.",
+            next,
+          });
+
+        res.status(200).json(cajero);
+      }
+    }
   }
 );
 
@@ -121,19 +155,31 @@ const getUsuarioCajero = asyncHandler(
     res: Response,
     next: NextFunction
   ): Promise<void | Response> => {
-    const { _id } = req.params;
+    const estado = Estado.getInstance();
 
-    const cajero = await Cajero.findOne({ _id });
+    if (estado.getArriba()) {
+      next();
+    } else {
+      const estado = Estado.getInstance();
 
-    if (!cajero)
-      return notFound({
-        message: "No se halló ningún cajero con el id provisto.",
-        next,
-      });
+      if (estado.getArriba()) {
+        next();
+      } else {
+        const { _id } = req.params;
 
-    const usuario = await Usuario.findOne({ entidad_asociada: cajero._id });
+        const cajero = await Cajero.findOne({ _id });
 
-    res.status(200).json(usuario);
+        if (!cajero)
+          return notFound({
+            message: "No se halló ningún cajero con el id provisto.",
+            next,
+          });
+
+        const usuario = await Usuario.findOne({ entidad_asociada: cajero._id });
+
+        res.status(200).json(usuario);
+      }
+    }
   }
 );
 
@@ -143,9 +189,15 @@ const getAllCajeros = asyncHandler(
     res: Response,
     next: NextFunction
   ): Promise<void | Response> => {
-    const cajeros = await Cajero.find({});
+    const estado = Estado.getInstance();
 
-    res.status(200).json(cajeros);
+    if (estado.getArriba()) {
+      next();
+    } else {
+      const cajeros = await Cajero.find({});
+
+      res.status(200).json(cajeros);
+    }
   }
 );
 
@@ -155,43 +207,49 @@ const updateCajero = asyncHandler(
     res: Response,
     next: NextFunction
   ): Promise<void | Response> => {
-    const session: ClientSession = await startSession();
+    const estado = Estado.getInstance();
 
-    try {
-      session.startTransaction();
-      const { _id } = req.params;
+    if (estado.getArriba()) {
+      next();
+    } else {
+      const session: ClientSession = await startSession();
 
-      const { cedula, nombre, apellido, sucursal_id } = req.body;
+      try {
+        session.startTransaction();
+        const { _id } = req.params;
 
-      const cashierFound: any = await Cajero.findById(_id).session(session);
+        const { cedula, nombre, apellido, sucursal_id } = req.body;
 
-      if (!cashierFound)
-        return notFound({
-          message: "No se halló ningún cajero con el _id provisto.",
-          next,
+        const cashierFound: any = await Cajero.findById(_id).session(session);
+
+        if (!cashierFound)
+          return notFound({
+            message: "No se halló ningún cajero con el _id provisto.",
+            next,
+          });
+
+        const newFields = {
+          cedula: cedula || cashierFound.cedula,
+          nombre: nombre || cashierFound.nombre,
+          apellido: apellido || cashierFound.apellido,
+          sucursal: (sucursal_id as Types.ObjectId) || cashierFound.sucursal,
+        };
+
+        await Cajero.updateOne(cashierFound, newFields, { session });
+
+        await session.commitTransaction();
+        session.endSession();
+
+        res.status(200).json({
+          exito: true,
+          mensaje: "Cajero actualizado satisfactoriamente.",
         });
+      } catch (error) {
+        await session.abortTransaction();
+        session.endSession();
 
-      const newFields = {
-        cedula: cedula || cashierFound.cedula,
-        nombre: nombre || cashierFound.nombre,
-        apellido: apellido || cashierFound.apellido,
-        sucursal: (sucursal_id as Types.ObjectId) || cashierFound.sucursal,
-      };
-
-      await Cajero.updateOne(cashierFound, newFields, { session });
-
-      await session.commitTransaction();
-      session.endSession();
-
-      res.status(200).json({
-        exito: true,
-        mensaje: "Cajero actualizado satisfactoriamente.",
-      });
-    } catch (error) {
-      await session.abortTransaction();
-      session.endSession();
-
-      return errorHandler(error, req, res, next);
+        return errorHandler(error, req, res, next);
+      }
     }
   }
 );
@@ -202,57 +260,63 @@ const deleteCajero = asyncHandler(
     res: Response,
     next: NextFunction
   ): Promise<void | Response> => {
-    const session: ClientSession = await startSession();
+    const estado = Estado.getInstance();
 
-    try {
-      session.startTransaction();
-      const { _id } = req.params;
+    if (estado.getArriba()) {
+      next();
+    } else {
+      const session: ClientSession = await startSession();
 
-      const cajero: any = await Cajero.findById(_id).session(session);
+      try {
+        session.startTransaction();
+        const { _id } = req.params;
 
-      if (!cajero) {
-        return notFound({
-          message: "No se encontró ningún cajero con el _id provisto.",
-          next,
+        const cajero: any = await Cajero.findById(_id).session(session);
+
+        if (!cajero) {
+          return notFound({
+            message: "No se encontró ningún cajero con el _id provisto.",
+            next,
+          });
+        }
+
+        if (cajero.cuadres.length > 0) {
+          return next(
+            new ErrorResponse(
+              "No puede eliminar a este cajero porque tiene cuadres/operaciones asociadas.",
+              401
+            )
+          );
+        }
+
+        const sucursalAsociada: any = await Sucursal.findById(
+          cajero.sucursal
+        ).session(session);
+
+        const idxToDelete = sucursalAsociada.cajeros.indexOf(cajero._id);
+
+        sucursalAsociada.cajeros.splice(idxToDelete, 1);
+
+        await sucursalAsociada.save();
+
+        if (cajero.usuario)
+          await Usuario.deleteOne({ _id: cajero.usuario }, { session });
+
+        await Cajero.deleteOne({ _id: cajero._id }, { session });
+
+        await session.commitTransaction();
+
+        res.status(200).json({
+          exito: true,
+          mensaje: "!Cajero eliminado satisfactoriamente!",
         });
+      } catch (error) {
+        await session.abortTransaction();
+
+        return errorHandler(error, req, res, next);
+      } finally {
+        session.endSession();
       }
-
-      if (cajero.cuadres.length > 0) {
-        return next(
-          new ErrorResponse(
-            "No puede eliminar a este cajero porque tiene cuadres/operaciones asociadas.",
-            401
-          )
-        );
-      }
-
-      const sucursalAsociada: any = await Sucursal.findById(
-        cajero.sucursal
-      ).session(session);
-
-      const idxToDelete = sucursalAsociada.cajeros.indexOf(cajero._id);
-
-      sucursalAsociada.cajeros.splice(idxToDelete, 1);
-
-      await sucursalAsociada.save();
-
-      if (cajero.usuario)
-        await Usuario.deleteOne({ _id: cajero.usuario }, { session });
-
-      await Cajero.deleteOne({ _id: cajero._id }, { session });
-
-      await session.commitTransaction();
-
-      res.status(200).json({
-        exito: true,
-        mensaje: "!Cajero eliminado satisfactoriamente!",
-      });
-    } catch (error) {
-      await session.abortTransaction();
-
-      return errorHandler(error, req, res, next);
-    } finally {
-      session.endSession();
     }
   }
 );
@@ -263,87 +327,96 @@ const createUsuarioCajero = asyncHandler(
     res: Response,
     next: NextFunction
   ): Promise<void | Response> => {
-    const session: ClientSession = await startSession();
-    session.startTransaction();
+    const estado = Estado.getInstance();
 
-    try {
-      const {
-        cedula,
-        email,
-        contrasenia,
-        tipo_entidad_asociada,
-        perfil,
-      } = req.body;
+    if (estado.getArriba()) {
+      next();
+    } else {
+      const session: ClientSession = await startSession();
+      session.startTransaction();
 
-      const invalidFields: string = validateRegistration(req);
+      try {
+        const {
+          cedula,
+          email,
+          contrasenia,
+          tipo_entidad_asociada,
+          perfil,
+        } = req.body;
 
-      if (invalidFields) return next(new ErrorResponse(invalidFields, 400));
+        const invalidFields: string = validateRegistration(req);
 
-      // Check if there's already a user with that email
-      const userFound = await Usuario.findOne({ email }).session(session);
+        if (invalidFields) return next(new ErrorResponse(invalidFields, 400));
 
-      if (userFound)
-        return next(
-          new ErrorResponse(
-            "El correo electrónico provisto ya está tomado.",
-            400
-          )
+        // Check if there's already a user with that email
+        const userFound = await Usuario.findOne({ email }).session(session);
+
+        if (userFound)
+          return next(
+            new ErrorResponse(
+              "El correo electrónico provisto ya está tomado.",
+              400
+            )
+          );
+
+        if (tipo_entidad_asociada !== "Cajero")
+          return next(
+            new ErrorResponse(
+              "Debe proveer 'Teller' como entidad asociada.",
+              400
+            )
+          );
+
+        const cajero: any = await Cajero.findOne({ cedula }).session(session);
+
+        if (!cajero)
+          return notFound({
+            message: "No se encontró ningún cajero con la cédula provista.",
+            next,
+          });
+
+        const usuarioRegistrado = await Usuario.findOne({
+          entidad_asociada: cajero._id,
+        }).session(session);
+
+        if (usuarioRegistrado)
+          return next(
+            new ErrorResponse("Usted ya posee una cuenta de cajero.", 400)
+          );
+
+        const perfilFound = await Perfil.findOne({ rol: perfil }).session(
+          session
         );
 
-      if (tipo_entidad_asociada !== "Cajero")
-        return next(
-          new ErrorResponse("Debe proveer 'Teller' como entidad asociada.", 400)
+        if (!perfilFound)
+          return notFound({ message: "El perfil provisto es inválido.", next });
+
+        const newUser: any = await Usuario.create(
+          [
+            {
+              email,
+              contrasenia,
+              tipo_entidad_asociada,
+              entidad_asociada: cajero._id,
+              perfil: perfilFound._id,
+            },
+          ],
+          { session }
         );
 
-      const cajero: any = await Cajero.findOne({ cedula }).session(session);
+        cajero.usuario = newUser[0]._id;
+        await cajero.save();
 
-      if (!cajero)
-        return notFound({
-          message: "No se encontró ningún cajero con la cédula provista.",
-          next,
-        });
+        await session.commitTransaction();
 
-      const usuarioRegistrado = await Usuario.findOne({
-        entidad_asociada: cajero._id,
-      }).session(session);
+        sendTokenResponse(newUser[0], 201, res, "sign up");
+      } catch (error) {
+        await session.abortTransaction();
 
-      if (usuarioRegistrado)
-        return next(
-          new ErrorResponse("Usted ya posee una cuenta de cajero.", 400)
-        );
-
-      const perfilFound = await Perfil.findOne({ rol: perfil }).session(
-        session
-      );
-
-      if (!perfilFound)
-        return notFound({ message: "El perfil provisto es inválido.", next });
-
-      const newUser: any = await Usuario.create(
-        [
-          {
-            email,
-            contrasenia,
-            tipo_entidad_asociada,
-            entidad_asociada: cajero._id,
-            perfil: perfilFound._id,
-          },
-        ],
-        { session }
-      );
-
-      cajero.usuario = newUser[0]._id;
-      await cajero.save();
-
-      await session.commitTransaction();
-
-      sendTokenResponse(newUser[0], 201, res, "sign up");
-    } catch (error) {
-      await session.abortTransaction();
-
-      return errorHandler(error, req, res, next);
-    } finally {
-      session.endSession();
+        return errorHandler(error, req, res, next);
+      } finally {
+        session.endSession();
+      }
     }
   }
 );
